Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/xhr/send-sync-response-event-order-expected.txt


FAIL XMLHttpRequest: The send() method: event order when synchronous flag is set assert_equals: expected "load(12,12,true)" but got "load(0,0,false)"
